requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlmodel>=0.0.8
python-dotenv>=0.19.0
pydantic>=1.9.0
sqlalchemy>=1.4.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
Pillow>=10.0.0
streamlit>=1.2.0
plotly>=5.3.1
pandas>=1.3.3
openai>=1.35.0
pytest>=6.2.5
httpx>=0.18.2 
requests-oauthlib>=1.3.0
authlib>=1.0.0
scikit-learn==1.4.0
spacy==3.7.2
numpy==1.26.3 